Implement changes from issue #5 (#7)
* Increase recv buffer size to UDP datagram max data size

* Add OOOI fields to acarsGenerator

* Fix syntax error (indenting)

* Fix syntax error (variable name)

* Add curl, ncat & net-tools

* Rename & touch JSON log files

* Change splitters to servers

* Update feeders to pull data from servers

* Update verbose logging

* Update webapp to pull data from servers

* Add startup checks

* Add stats services

* Linting

* Change UDP-SENDTO to UDP

* Keep last 2 hours of msgs

* change filename format

* Catching invalid JSON

* Implement a healthcheck

* Updates to healthchecks

* QUIET_LOGS implementation

* Linting

* Linting

* Updating linter to remove SC2009 globally

Not sure this is the best way, but I don't see why that error is a concern...

* Reverting, using pgrep

* Syntax error

* Making acars_stats strict pattern matching

* Removing VERBOSE ENV variable/healthcheck.sh update

VERBOSE ENV variable removed, replaced with ENABLE_WEB and QUIET_LOGS to clarify

* Attempted fix for large message JSON truncation

* Fix for ACARS small packet truncation

* The whole fix...

* Update run

* Updating debug levels, attempted fix for client disconnect issue

Added in EXTREME_DEBUG ENV variable to mute some debug messages

Added in socket.io error handlers

I believe the issue I'm seeing with client disconnect is because the server gets overloaded and doesn't process the client ping/pong interaction quick enough, and then gunicorn marks the session as disconnected. Then, the client sends another ping later and causes the traceback because the session is disconnected. I'm having trouble with fully wrapping my mind around this issue, but the google suggests increasing the timeout might alleviate the issue, and I think I've got it in the right spot. Testing now.

* Linting fix

* Update application.js

* Add script for device selection

* Fix device selection

* fix flask warning

* Fix typo

* Readme updates

Co-authored-by: fredclausen <clausen.fred@icloud.com>