Web client improvement (#14)
* Increase recv buffer size to UDP datagram max data size

* Add OOOI fields to acarsGenerator

* Fix syntax error (indenting)

* Fix syntax error (variable name)

* Add curl, ncat & net-tools

* Rename & touch JSON log files

* Change splitters to servers

* Update feeders to pull data from servers

* Update verbose logging

* Update webapp to pull data from servers

* Add startup checks

* Add stats services

* Linting

* Change UDP-SENDTO to UDP

* Keep last 2 hours of msgs

* change filename format

* Catching invalid JSON

* Implement a healthcheck

* Updates to healthchecks

* QUIET_LOGS implementation

* Linting

* Linting

* Updating linter to remove SC2009 globally

Not sure this is the best way, but I don't see why that error is a concern...

* Reverting, using pgrep

* Syntax error

* Making acars_stats strict pattern matching

* Removing VERBOSE ENV variable/healthcheck.sh update

VERBOSE ENV variable removed, replaced with ENABLE_WEB and QUIET_LOGS to clarify

* Attempted fix for large message JSON truncation

* Fix for ACARS small packet truncation

* The whole fix...

* Update run

* Updating debug levels, attempted fix for client disconnect issue

Added in EXTREME_DEBUG ENV variable to mute some debug messages

Added in socket.io error handlers

I believe the issue I'm seeing with client disconnect is because the server gets overloaded and doesn't process the client ping/pong interaction quick enough, and then gunicorn marks the session as disconnected. Then, the client sends another ping later and causes the traceback because the session is disconnected. I'm having trouble with fully wrapping my mind around this issue, but the google suggests increasing the timeout might alleviate the issue, and I think I've got it in the right spot. Testing now.

* Linting fix

* Update application.js

* Add async_handlers=True to SocketIO

* Generate the local docker file on the fly

* Websockets work

* Move ACARS decoding to new thread

* VDLM listener in own thread

* Linting

* db initial structure

* Database initing and write function

* DB write in own thread

* Removing useless line

* DB writing works!

Also included a spammer program for spamming the program with messages for testing

* Update acarshub_db.py

* Cleaning up test print messages

* Cleaned up db key testing/slow down db listener thread

* adding in missing 'data' key

* README.md update for database retention

* Threading improvements

Moved html generation to own function, cleaned up imports in spammer

* removing debug statement

* database pruning on schedule

* Create TODO.md

* HTML updates

* scheduler function call fixup

* Missing key libacars

* IATA to ICAO airline line lookup

Data source is bogus, so the callsign may or may not be accurate, and the generated flight aware link might be bad as a result. Will need a better source.

* Covering off error condition if IATA code isn't in the DB

Probably need to have a better think on this, as the generated FA link will not work with an IATA code

* Linting

* Spammer updates, minor web page updates, db logging messages

* Website CSS / Tooltip start

* Removing tool-tip/link if IATA to ICAO code conversion fails

* Linting....

* Explaining IATA/ICAO conversion errors

* Catching DB exceptions

Also updated generated HTML code

* Linting

* Basic structure for searching database

Doesn't actually search the terms yet, but it does query the db for random data and returns/displays it

* Database query, split in to pages

Not perfect yet. Display is off, but it does (mostly) what it should

* Search cleaned up

* Comments

* Pause live updates

* CSS/Cleanup/client connction tracking

* linting

* Small bug fixes/cleanup/html output change

* Cleaning up page generation logic

* Run search after user changes dropobox selection

* broadcast search results to the intended client and not everyone

* First go of stats

* images on the website

Kinda broken - browser caches the image and doesn't update

* attempt to bypass browser caching

* HTTP header cache time to 0

Also reducing the size of the generated image

* Fixing six month image

* stat image auto-update

* Adding data in to rrd via python

* Deleting db init shell script

* Moved completely over to python for stats

* Linting

* Increasing stat web refresh, linting function name

* UI Improvements

* Bug fix

A hard to notice bug, but now if the que is > 1 message long the oldest message is grabbed

* Linting

* Multi-receive error condition handled

* Bug fix for previous commit

* libacars fixes

* test namespace to main namespace

* Spammer improvements, altering on load behavior

* sdr listener to a single function

* Add in Errors in RRD

* RRD Color improvements?

* Delete Dockerfile.local

* Update .gitignore

* Add action to cancel duplicate workflows

* MD Updates

* add rrdtool & librrd-dev to runner

* Use sudo to install packages

* Linting

Co-authored-by: Mike Nye <mike.nye@gmail.com>
Co-authored-by: fredclausen <clausen.fred@icloud.com>