Try/Except blocks around all emit statements in on connect
With the new dumpvdl2 in testing I managed to get some bad data in to the database and it caused issues with types and other things during a load event, which would cause everything to grind to a halt. Catching each step of the load event and not letting it kill execution will prevent this.