Cache bust dev fixes
All builds should use the prod target, unless building and running ACARS Hub outside of the container. dev target does not rewrite the JS and other asset names to help with caching issues. We want that outside of the container because a simple page refresh allows us to get the updated JS builds behind the scenes because nothing is cached by default if acarshub handles the web server. nginx, inside the container, does send cached header.