Improve readability in the update_keys function
We already remove "stale" (aka None and "") keys when processing the message. After doing so we only check if the key exists, because if it does it actually should hold data.

In addition, now we catch any potential errors converting the input decimal number to hex, just in case the data in malformed.