Chrome (mostly) browser performance issues
In chrome if a tab was backgrounded for more than 3-5 minutes it would stop processing messages and presumably hold them in a queue. Once the user brought the task to the foreground it would take forever to process and the page was unresponsive. Refreshing didn't do anything.

Additionally, the DOM tree for messages (which is overly complex, needs to be fixed) was being dumped and de-done every single time a message came in. CPU usage is much, much too high and chrome in particular doesn't seem to like that.

This fix attempts to fix the problem by doing the following:

* Change limit of 50 message groups displayed on the Live Message page to 50
* When a message comes in only change the part of the DOM tree that was affected. IE, if a message is matched delete that message group from the DOM and promote it back to the top. If a message is new, prepend it to the top of the tree. To expire old messages check the number of div elements in the #log div and remove last child if more than the alloted number.
* Handle being backgrounded by telling all pages to stop their update cycles by toggling everything "off". New messages should be processed, still, but the DOM is never walked.